+++ src/admin_ui/src/admin/widgets/mapping_input.tsx
@@ -47,5 +47,3 @@ export const MappingInput: FC<MappingInputProps> =
             </FormControl>
         );
     }
-
-export default MappingInput;